refactor: split agent buttons, implement updated services hooks
