+++ frontend/components/MainPage/header/AgentButton/AgentNotRunningButton.tsx
@@ -1,14 +1,13 @@
-import { InfoCircleOutlined } from '@ant-design/icons';
-import { Button, ButtonProps, Flex, Popover, Tooltip, Typography } from 'antd';
+import { Button, ButtonProps } from 'antd';
 import { useCallback, useMemo } from 'react';
 
 import { MiddlewareChain, MiddlewareDeploymentStatus } from '@/client';
-import { COLOR } from '@/constants/colors';
+import { STAKING_PROGRAMS } from '@/config/stakingPrograms';
 import { DEFAULT_STAKING_PROGRAM_ID } from '@/context/StakingProgramProvider';
 import { StakingProgramId } from '@/enums/StakingProgram';
 import { useBalance } from '@/hooks/useBalance';
 import { useElectronApi } from '@/hooks/useElectronApi';
-import { useReward } from '@/hooks/useReward';
+import { useService } from '@/hooks/useService';
 import { useServices } from '@/hooks/useServices';
 import { useServiceTemplates } from '@/hooks/useServiceTemplates';
 import { useStakingContractInfo } from '@/hooks/useStakingContractInfo';
@@ -19,116 +18,24 @@ import { ServicesService } from '@/service/Services';
 import { WalletService } from '@/service/Wallet';
 import { delayInSeconds } from '@/utils/delay';
 
-import {
-  CannotStartAgentDueToUnexpectedError,
-  CannotStartAgentPopover,
-} from './CannotStartAgentPopover';
-import { requiredGas } from './constants';
-import { LastTransaction } from './LastTransaction';
-
-const { Text, Paragraph } = Typography;
-
-const LOADING_MESSAGE =
-  "Starting the agent may take a while, so feel free to minimize the app. We'll notify you once it's running. Please, don't quit the app.";
-
-const IdleTooltip = () => (
-  <Tooltip
-    placement="bottom"
-    arrow={false}
-    title={
-      <Paragraph className="text-sm m-0">
-        Your agent earned rewards for this epoch, so decided to stop working
-        until the next epoch.
-      </Paragraph>
-    }
-  >
-    <InfoCircleOutlined />
-  </Tooltip>
-);
-
-const AgentStartingButton = () => (
-  <Popover
-    trigger={['hover', 'click']}
-    placement="bottomLeft"
-    showArrow={false}
-    content={
-      <Flex vertical={false} gap={8} style={{ maxWidth: 260 }}>
-        <Text>
-          <InfoCircleOutlined style={{ color: COLOR.BLUE }} />
-        </Text>
-        <Text>{LOADING_MESSAGE}</Text>
-      </Flex>
-    }
-  >
-    <Button type="default" size="large" ghost disabled loading>
-      Starting...
-    </Button>
-  </Popover>
-);
-
-const AgentStoppingButton = () => (
-  <Button type="default" size="large" ghost disabled loading>
-    Stopping...
-  </Button>
-);
-
-const AgentRunningButton = () => {
-  const { showNotification } = useElectronApi();
-  const { isEligibleForRewards } = useReward();
-  const { service, setIsServicePollingPaused, setServiceStatus } =
-    useServices();
-
-  const handlePause = useCallback(async () => {
-    if (!service) return;
-    // Paused to stop overlapping service poll while waiting for response
-    setIsServicePollingPaused(true);
-
-    // Optimistically update service status
-    setServiceStatus(MiddlewareDeploymentStatus.STOPPING);
-    try {
-      await ServicesService.stopDeployment(service.hash);
-    } catch (error) {
-      console.error(error);
-      showNotification?.('Error while stopping agent');
-    } finally {
-      // Resume polling, will update to correct status regardless of success
-      setIsServicePollingPaused(false);
-    }
-  }, [service, setIsServicePollingPaused, setServiceStatus, showNotification]);
-
-  return (
-    <Flex gap={10} align="center">
-      <Button type="default" size="large" onClick={handlePause}>
-        Pause
-      </Button>
-
-      <Flex vertical>
-        {isEligibleForRewards ? (
-          <Text type="secondary" className="text-sm">
-            Agent is idle&nbsp;
-            <IdleTooltip />
-          </Text>
-        ) : (
-          <Text type="secondary" className="text-sm loading-ellipses">
-            Agent is working
-          </Text>
-        )}
-        <LastTransaction />
-      </Flex>
-    </Flex>
-  );
-};
+import { requiredGas } from '../constants';
 
 /** Button used to start / deploy the agent */
-const AgentNotRunningButton = () => {
+export const AgentNotRunningButton = () => {
   const { wallets, masterSafeAddress } = useWallet();
+
   const {
-    service,
-    serviceStatus,
-    setServiceStatus,
-    setIsServicePollingPaused,
-    updateServicesState,
+    selectedService,
+    setPaused: setIsServicePollingPaused,
+    isLoaded,
+    refetch: updateServicesState,
   } = useServices();
+
+  const { service, deploymentStatus, setDeploymentStatus } = useService({
+    serviceConfigId:
+      isLoaded && selectedService ? selectedService?.service_config_id : '',
+  });
+
   const { serviceTemplate } = useServiceTemplates();
   const { showNotification } = useElectronApi();
   const {
@@ -146,16 +53,15 @@ const AgentNotRunningButton = () => {
     setIsPaused: setIsStakingContractInfoPollingPaused,
     updateActiveStakingContractInfo,
   } = useStakingContractInfo();
+
   const { activeStakingProgramId } = useStakingProgram();
 
   // const minStakingDeposit =
   //   stakingContractInfoRecord?.[activeStakingProgram ?? defaultStakingProgram]
   //     ?.minStakingDeposit;
 
-  const requiredOlas = getMinimumStakedAmountRequired(
-    serviceTemplate,
-    activeStakingProgramId ?? DEFAULT_STAKING_PROGRAM_ID,
-  );
+  const requiredOlas =
+    STAKING_PROGRAMS[activeStakingProgramId]?.minStakingDeposit; // TODO: fix activeStakingProgramId
 
   const safeOlasBalance = safeBalance?.OLAS;
   const safeOlasBalanceWithStaked =
@@ -177,7 +83,7 @@ const AgentNotRunningButton = () => {
     setIsStakingContractInfoPollingPaused(true);
 
     // Mock "DEPLOYING" status (service polling will update this once resumed)
-    setServiceStatus(MiddlewareDeploymentStatus.DEPLOYING);
+    setDeploymentStatus(MiddlewareDeploymentStatus.DEPLOYING);
 
     // Get the active staking program id; default id if there's no agent yet
     const stakingProgramId: StakingProgramId =
@@ -190,7 +96,7 @@ const AgentNotRunningButton = () => {
       }
     } catch (error) {
       console.error(error);
-      setServiceStatus(undefined);
+      setDeploymentStatus(undefined);
       showNotification?.('Error while creating safe');
       setIsStakingContractInfoPollingPaused(false);
       setIsServicePollingPaused(false);
@@ -208,7 +114,7 @@ const AgentNotRunningButton = () => {
       });
     } catch (error) {
       console.error(error);
-      setServiceStatus(undefined);
+      setDeploymentStatus(undefined);
       showNotification?.('Error while deploying service');
       setIsServicePollingPaused(false);
       setIsBalancePollingPaused(false);
@@ -225,7 +131,7 @@ const AgentNotRunningButton = () => {
     }
 
     // Can assume successful deployment
-    setServiceStatus(MiddlewareDeploymentStatus.DEPLOYED);
+    setDeploymentStatus(MiddlewareDeploymentStatus.DEPLOYED);
 
     // TODO: remove this workaround, middleware should respond when agent is staked & confirmed running after `createService` call
     await delayInSeconds(5);
@@ -233,7 +139,7 @@ const AgentNotRunningButton = () => {
     // update provider states sequentially
     // service id is required before activeStakingContractInfo & balances can be updated
     try {
-      await updateServicesState(); // reload the available services
+      await updateServicesState?.(); // reload the available services
       await updateActiveStakingContractInfo(); // reload active staking contract with new service
       await updateBalances(); // reload the balances
     } catch (error) {
@@ -249,7 +155,7 @@ const AgentNotRunningButton = () => {
     setIsServicePollingPaused,
     setIsBalancePollingPaused,
     setIsStakingContractInfoPollingPaused,
-    setServiceStatus,
+    setDeploymentStatus,
     masterSafeAddress,
     showNotification,
     activeStakingProgramId,
@@ -263,15 +169,15 @@ const AgentNotRunningButton = () => {
     // if the agent is NOT running and the balance is too low,
     // user should not be able to start the agent
     const isServiceInactive =
-      serviceStatus === MiddlewareDeploymentStatus.BUILT ||
-      serviceStatus === MiddlewareDeploymentStatus.STOPPED;
+      deploymentStatus === MiddlewareDeploymentStatus.BUILT ||
+      deploymentStatus === MiddlewareDeploymentStatus.STOPPED;
     if (isServiceInactive && isLowBalance) {
       return false;
     }
 
-    if (serviceStatus === MiddlewareDeploymentStatus.DEPLOYED) return false;
-    if (serviceStatus === MiddlewareDeploymentStatus.DEPLOYING) return false;
-    if (serviceStatus === MiddlewareDeploymentStatus.STOPPING) return false;
+    if (deploymentStatus === MiddlewareDeploymentStatus.DEPLOYED) return false;
+    if (deploymentStatus === MiddlewareDeploymentStatus.DEPLOYING) return false;
+    if (deploymentStatus === MiddlewareDeploymentStatus.STOPPING) return false;
 
     if (!requiredOlas) return false;
 
@@ -290,7 +196,7 @@ const AgentNotRunningButton = () => {
 
     return hasEnoughOlas && hasEnoughEth;
   }, [
-    serviceStatus,
+    deploymentStatus,
     service,
     storeState?.isInitialFunded,
     isEligibleForStaking,
@@ -311,48 +217,4 @@ const AgentNotRunningButton = () => {
   const buttonText = `Start agent ${service ? '' : '& stake'}`;
 
   return <Button {...buttonProps}>{buttonText}</Button>;
-};
-
-export const AgentButton = () => {
-  const { service, serviceStatus, hasInitialLoaded } = useServices();
-  const { isEligibleForStaking, isAgentEvicted } = useStakingContractInfo();
-
-  return useMemo(() => {
-    if (!hasInitialLoaded) {
-      return <Button type="primary" size="large" disabled loading />;
-    }
-
-    if (serviceStatus === MiddlewareDeploymentStatus.STOPPING) {
-      return <AgentStoppingButton />;
-    }
-
-    if (serviceStatus === MiddlewareDeploymentStatus.DEPLOYING) {
-      return <AgentStartingButton />;
-    }
-
-    if (serviceStatus === MiddlewareDeploymentStatus.DEPLOYED) {
-      return <AgentRunningButton />;
-    }
-
-    if (!isEligibleForStaking && isAgentEvicted)
-      return <CannotStartAgentPopover />;
-
-    if (
-      !service ||
-      serviceStatus === MiddlewareDeploymentStatus.STOPPED ||
-      serviceStatus === MiddlewareDeploymentStatus.CREATED ||
-      serviceStatus === MiddlewareDeploymentStatus.BUILT ||
-      serviceStatus === MiddlewareDeploymentStatus.DELETED
-    ) {
-      return <AgentNotRunningButton />;
-    }
-
-    return <CannotStartAgentDueToUnexpectedError />;
-  }, [
-    hasInitialLoaded,
-    serviceStatus,
-    isEligibleForStaking,
-    isAgentEvicted,
-    service,
-  ]);
 };
