+++ frontend/components/MainPage/header/AgentButton/AgentStoppingButton.tsx
@@ -1,0 +1,7 @@
+import { Button } from 'antd';
+
+export const AgentStoppingButton = () => (
+  <Button type="default" size="large" ghost disabled loading>
+    Stopping...
+  </Button>
+);
